# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is example for front end",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yurayur1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vatanparvarbackend.napaautomotive.uz/",
            "description": "This is domen for frontend"
        }
    ],
    "paths": {
...
        "/api/tagsearch/{search}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get data by short name",
                "operationId": "TagSearch",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Izlash so'zini kiriting",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "text"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/titlesearch/{search}": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get data by short name",
                "operationId": "TitleSearch",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Izlash so'zini kiriting",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "text"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/sendProject": {
            "post": {
                "tags": [
                    "Send Project"
                ],
                "summary": "Get data by short name",
                "operationId": "sendproject",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "required|max:50",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "required|max:100",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "required|min:2|max:15",
                                        "type": "string"
                                    },
                                    "columnName": {
                                        "description": "required|max:100",
                                        "type": "string"
                                    },
                                    "about": {
                                        "description": "required",
                                        "type": "string"
                                    },
                                    "images": {
                                        "description": "required",
                                        "type": "file"
                                    },
                                    "socials": {
                                        "description": "required",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}